requirements.txt
telethon
python-dotenv
PyQt6
